requirements.txt
solc==0.0.0a0
py-solc-x==1.1.1
aiohttp==3.10.11
python-dotenv==1.0.1
web3==7.6.0
asyncio==3.4.3
web3-ethereum-defi==0.9
async-timeout==5.0.1
